requirements.txt.txt
colorama==0.4.6
Nuitka==2.6.9
ordered-set==4.1.0
pillow==11.1.0
qrcode==8.0
zstandard==0.23.0
